refactor: extract table and chart utilities
